Github action configuration to publish npm package
The action is publishing contracts to npm package repository on each
push to master which updates files in `contracts` directory.
After publishing the package it bumps up version of the npm package to a
next prerelease. The update to `package.json` file is commited to te
repository with Heimdall user. The commit message has `[CI]` prefix
which helps to filter out such commits from next action runs.