API for selecting group with unique members and a temporary implementation
selectSetGroup selects a new group of operators of the provided size
based on the provided pseudo-random seed. All operators in the selected
group are unique. If there are not enough operators registered in the
pool, function reverts.

In this commit we ship an API as well as a temporary implementation -
working but insecure. This code will be replaced.