Add comments tracking the arbitrarily named variables foo and bar
The actual meanings of the variables `foo` and `bar` are tracked in the comments
because of the limitations of the Solidity compiler.
I figured trying to give them meaningful names would just be confusing.
Now they're as stupid and annoying as the reason why I had to do this.
This will definitely not bite us in the ass later.