Fix the return values on pickWeightedSlot() in Branch.sol
Linting had changed pickWeightedSlot() at Branch.sol
so that the second return value was no longer the mutated weight,
but the input weight.

This completely broke everything
but only the codeshoggoth that is selectSetGroup() in BondedSortitionPool.sol
actually caught it,
so I naturally thought
the problem was with my horrible abomination of blobcode
and not the utility functions that were just about perfect the way I left them.

The incorrectness would've been caught by the require()
that I had removed for performance reasons
because the function was self-evidently correct
when called with the correct inputs,
and it made more sense to enforce input correctness once
at the beginning (so outside pickWeightedSlot())
instead of redundantly inside the loop.

The test cases in sortitionTest.js failed to catch this
because the sortition ended up hitting the correct leaves
even with incorrect behavior.

The naive method selectSetGroupB() concealed this problem
because it just tried again until it found its unique members,
and our tests weren't large enough
(more than 16 operators and a group size of >16
would've inevitably revealed the problem)

The codeshoggoth caught the incorrectness
because it was relying on the output of the final call to pickWeightedSlot()
inside pickWeightedLeafWithIndex()
to return the difference between the starting index of the selected leaf
and the originally queried index,
but what it actually received was the original index unchanged
(as it had passed intact through the loop).
When the original index was subtracted from the original index,
the result was always zero
as the supposed starting index of the selected operator.

As a result of this, the codeshoggoth,
which actually tracked the starting indices of the selected operators,
only ever added 0, the starting index of the first operator,
to the list of operators to skip.
The index was associated with the weights of the previously selected operators,
resulting in it incorrectly skipping the first operator/s
(depending on how the weights summed up)
and getting duplicates of the others.