Add initial minimum bond as a bonded pool constructor argument
Bonded sortition pools must be able to accommodate variable bond amounts,
but a minimum bond value needs to be set and stored.
We could use the last `select(Set)Group()` call's bond value
as the subsequent minimum bond,
but this leaves the initial minimum bond unset.

An insufficient minimum bond enables a particular form of griefing
where a large number of operators authorize the pool for bonding
but only have a very small amount of bondable value available.
Then when the pool actually tries to select a group,
it hits a large number of ineligible operators whose bondable value is too low,
causing high and unpredictable gas costs.
The griefing is mildly difficult to pull off in any useful way
(requiring temporal arbitrage of gas prices,
losing out on opportunities to earn revenue
from other applications using the same bonding contract, and so on)
but it is still a way to harm applications.