Store the creation block in previously unused leaf space
We may need to use the creation block
to set an initialiation period for operators who join the pool
so they automatically become eligible after a certain number of blocks.
This helps mitigate "pool bumping" attacks
where a miner inserts or removes operators from the pool
to manipulate the composition of selected groups.