Enforce minimum stake requirements in pool constructor
In the least convenient world,
16 operators each have 2^15 weight,
and must all be inserted into separate trunks.
In this case, any trunk can still accept another 2^15 - 1 weight.
In the least convenient world,
any stake not held by those 16 operators is held by a single operator,
whom we can accommodate if their stake is at most 2^15 - 1.
This gives us a total pool capacity of 2^15 * 17 - 1;
as long as there can be at most that much staker weight in the world
the pool will be able to accommodate any non-overflowing operator.

From this information and the total supply of KEEP tokens,
we can derive the number of tokens the minimum stake must exceed.
This is enforced in the AbstractSortitionPool constructor
which is inherited by SortitionPool and BondedSortitionPool.