Change from 16-ary tree of uint16 to 8-ary tree of uint32
At the cost of approximately 10% gas,
the capacity of the tree is increased to 2 million operators
and all the finicky special-casing with trunks and weight overflows
is completely eliminated.

This breaks most internal tests in a way which is annoying to rewrite
so those parts have been commented out.
High-level user-visible logic works though,
and in theory this would be a drop-in replacement for 16-ary sortition pools
as far as the interface is concerned.