Make operators wait before they can be selected after joining
If operators can join the pool while the selection seed is known
(e.g. due to miner front-running or a public cached seed in the factory)
the outcome of selection can be manipulated.
By having a sufficiently high `INIT_BLOCKS` this can be prevented.

The current value of 10 is not sufficient to properly prevent attacks
but was chosen for ease of testing.
`INIT_BLOCKS` must be updated before mainnet.

The current method for skipping recently joined operators is rudimentary
and can lead to infinite loops if all operators have joined recently
or out-of-gas if too large a fraction of the pool weight has joined recently.
The technique `selectSetGroup()` uses to skip selected operators
can also be used to skip recent operators after first encountered,
as soon as I figure out
a good way to do dynamic-length in-memory arrays in Solidity.