Remove unnecessary parameters from Operator library
The `operator`, `outOfDate` and `leafPosition` fields are only useful
if we're doing batched operator deletion at the end.

Batched operator deletion would permit caching intermediate nodes in the tree,
instead of just root, for additional gas savings.
However, this would be offset
by the need to keep outdated operators in the list of skipped leaves.

Overall it doesn't seem worth the introduced complexity.