Add reseed function RNG and reseed when we select a group member
RNG.generateNewIndex() doesn't necessarily update the seed
so it can end up trying again with the same truncated index.
When an operator is deleted or skipped
the next one that gets picked is the operator that would have been selected
if the previous one hadn't been in the pool.
However, when an operator is selected for the group
not reseeding means
that the next selected operator will be the same one if allowed,
or the next one if duplicates are prohibited.
This is not the desired behavior.