Wrap Decision in a Fate struct to accommodate leaf updates
For sortition pools to deal with fully backed staking,
we need to be able to update the weights of selected operators on the fly.

Solidity doesn't support sum types
so we can't put the update weight in the Decision datatype
without some annoying bit twiddling.
Therefore, wrap Decision and optional weight in a Fate struct.