Minimum bondable value for Fully Backed Sortition Pool
The pool was updating the minimum bondable value for each
selectSetGroup call. The idea was to automatically remove from the pool
operators that are no longer able to satisfy application needs about the
minimum available bondable value and prevent operators with too low value
griefing signer selection.

We now have two minimum bond values: one defining the minimum unbonded
value the operator needs to have so that it can join and stay in the pool,
and another defining the required bond value for the current selection.
If the given operator has enough unbonded value to stay in the pool but it
does not have enough unbonded value for the current group selection, it is
skipped.

We assume a reasonable minimum bondable value is set on the pool
creation and it can be later updated by the pool owner (application).
We move the responsibility of keeping this value sane to the application.
If the application defines allowed lot sizes, just like tBTC, the application
may set the minimum bondable value to the minimum lot size and later skip
operators that are not eligible to satisfy bond for the current selection
instead of removing them from the pool. All operators without at least the
minimum unbonded value are removed from the pool during the group selection.

The change was originally implememnted to BondedSortitionPool in 46d187a.