Added bonding contract interface for fully backed bonding
We created a dedicated interface for the fully backed bonding as there
will be specific rules to verify for the group selection.
For regular BondedSortitionPool we had a staking contract reference
holding informations like eligibleStake, here we don't use staking
contract but all required informations are received from bonding
contract.