Use WEIGHT_WIDTH to index full weight range in RNG
The RNG was using the POSITION_BITS to determine how many total bits can be
used in a pool index. This limited the RNG's range to the total number
of positions in the pool, rather than the total weight the pool could
manage. In previous usage, this was sufficient for total weight as well,
but since the system has since moved to handling ETH fully backed pools
and denominate KEEP stakes by individual tokens instead of minstakes,
the ~2M position limit is no longer sufficient to represent the
magnitude of total weights that need to be handled.

The RNG now bases the total weight it can index into on the total weight
that the underlying sortition tree can manage, which is significantly
higher (~4B vs ~2M), represented by the WEIGHT_WIDTH constant.