Add rudimentary buidler setup
This allows the usage of buidler-vm instead of ganache for testing,
which in turn allows the usage of console.log in contracts.

A special script, buidler-vm:log, filters the buidler VM output to only
log output. This is useful for testing with logs.

Lastly, with ganache gone, its script is updated to note that Ganache
is no longer used, and the Solidity test workflow is updated to use
npm run buidler-vm instead of ganache.