Added operators banning to FullyBackedSortitionPool
Added a functionality of operators banning in FullyBackedSortitionPool
contract.

A pool owner can call a function to ban an operator. Once an operator gets
banned it is removed from the pool. The operator won't get selected to
new groups.

Operator can be banned even if it has never been registered or is not currently
registered in the pool.

A banned operator cannot join a pool.