Publish NPM package without commiting package version
Here we update script for publishing NPM package in GitHub Actions job.

For `master` branch we want to publish `pre` package on each `master`
merge. But there is no sense in doing so if nothing has changed in the
solidity contracts. Circle CI doesn't allow us to filter on changed
content to run a job, that's why we use GitHub Actions. We will publish
new `pre` package to NPM registry on master merge only if solidity
contracts have changed.

Publishing to NPM registry requires a `NPM_TOKEN` token to be configured
in GH project's secrets variables.