Use `-dev` suffix in the versioning of the `sortition-pools`
`-dev` suffix will be used for the versioning of the `sortition-pools`
package built with the purpose of being used on development environment.
Packages with the `-dev`-suffix will be pushed to the NPM registry by
the `NPM` GH Actions workflow.
Note that packages ment for the test environment (Ropsten) will be
versioned independently, based on the suffix provided during manual
triggering of the `Main` workflow in `ci` project.