Remove redundancy in leaf weight encoding
We need to store member ID numbers in the leaves for cheaper sortition.
For this, we need to free up an uint32 in the leaves.
Fortunately the weight of each operator was previously recorded twice,
once in the leaf and once in the branch above it.
By removing all reliance on the weight information in the leaf
we can repurpose that slot for the ID number.