Store IDs instead of weights in leaves and return them in selection
Complete the replacement of weights with ID numbers
and modify generalizedSelectGroup() to return the selected leaves.
Specific implementations of the abstract sortition pool
can filter the returned array to return what information they want.

This has a generally low impact on gas costs.
Group selections that involve updating leaves are a bit cheaper
as the actual leaf data is not modified, only the tree contents.
Large group selections are slightly more expensive.
Joining the pool is significantly more expensive on the first time,
but could be further optimised to some degree.