Represent operator ID as uint32
Leaf.make expects ID fits in 32 bits but we were not asserting anywhere
if this is really the case. From now on, when new operator joins the
pool we check if the max capacity was not exceeded.

Additionally, `getOperatorID` and `getIDOperator` functions operate now
on `uint32` officially. This is how the external code stores group
members data and it makes sense to have public sortition pool API
reflecting it, even though internally, for gas efficiency purposes,
sortition pool can work on uint256 IDs.