Allow to fetch operator addresses by their IDs in one call
Sometimes it is needed to fetch addresses of all operators in the group
based on their IDs. One example is random beacon DKG result validation
code that needs to to do ecrocover on received signatures. To avoid
doing external contract call in a loop, getIDOperators allows to fetch
addresses in just one call.