Add basic pool reward functionality
Rewards are allocated proportionally to operators present in the pool at payout
based on their weight in the pool.

To facilitate this, we use a global accumulator value to track the total rewards
one unit of weight would've earned since the creation of the pool.

Whenever a reward is paid, the accumulator is increased by the size of the
reward divided by the total weight of all eligible operators in the pool.

Each operator has an individual accumulator value, set to equal the global
accumulator when the operator joins the pool. This accumulator reflects the
amount of rewards that have already been accounted for with that operator.

Whenever an operator's weight in the pool changes, we can update the amount of
rewards the operator has earned by subtracting the operator's accumulator from
the global accumulator. This gives us the amount of rewards one unit of weight
has earned since the last time the operator's rewards have been updated. Then we
multiply that by the operator's previous (pre-change) weight to determine how
much rewards in total the operator has earned, and add this to the operator's
earned rewards. Finally, we set the operator's accumulator to the global
accumulator value.

In this commit the reward accumulator functionality is implemented and tested
separately of the sortition pool itself. Very basic mock integration to the pool
itself is also included.