Removed unused SortitionPool.minimumStake field
As a result of earlier changes, SortitionPool does no longer need
a reference to the staking contract and is no longer checking if
the given operator is eligible to join - this is a responsibility
of the application now. Minimum stake field in SortitionPool was
overlooked during these changes - cleaning up now.