Remove `DynamicArray`
`DynamicArray` is a clever bit of spooky assembly bit-twiddling, but it is
spooky assembly bit-twiddling. We don't need it with the new design where the
sortition pool is static during group selection and duplicate indices aren't
skipped, and it would be that much less code to audit and worry about.