Proxy withdrawRewards through the application
Say we are creating a new group every month and that the group lifetime
is 2 months. Say we have group `a` active and group `b` will soon be
created. Let's say that the operator requests authorization decrease at
`x` moment of time:

```
|---------a-----------|
   x      |---------b----------|
                     |---------c-----------|

```

If we do not remove operator from the sortition pool at `x`, it may happen
that the operator gets selected to group `b` and they might be able to
approve authorization decrease before `b` ends its life. This is not ideal
and it generally feels wrong not to remove the operator from the sortition
pool immediately if we know that operator will be unwinding their node.

If we remove the operator from the pool at `x` they stop earning rewards and
given that the authorization decrease delay needs to be long enough given
the long group lifetime, this situation is less than ideal.

We should rework rewards so that it is possible to remove someone from the
pool while still giving them a chance to earn rewards. This will be addressed
separately but it still does not solve all problems.

We need to find a way to have the operator call `approveAuthorizationDecrease`
function on the application contract as soon as their authorization decrease
delay ends to stop them from earning rewards for nothing.

One option, the one implemented in this PR, is to make
`SortitionPool.withdrawRewards` protected by `onlyOwner` and proxy it through
the application contract. This lets us eliminate `IStaking` from `SortitionPool`
and keep dependency graph clean: `TokenStaking <-> Application  <-> SortitionPool`.
Moreover, application will be able to reduce the amount of reward withdrawn
proportionally to the delay and send this amount back to the sortition pool or to
some governable reward funder account.