Proxy reward withdrawal through application contract
Previously, `SortitionPool.withdrawRewards` could be called directly and
the beneficiary address was fetched from the staking contract based on
the operator address.

This is no longer going to work - T `TokenStaking contract` does not store
operators but staking providers - operator is now a per-application role.

We now expect `SortitionPool.withdrawReward` call to be proxied through the
application (`RandomBeacon` or `WalletRegistry` contract) that will fetch
the beneficiary address from T `TokenStaking`.