Simplify the sortition tree test structure
Previously we had a lot of tests that had a context with a single `it`
expectation, where the bulk of the test setup was abstracted into a
`beforeEach`. This is useful when we want to run the code in
`beforeEach` for multiple different `it` contexts, but if that
abstraction isn't being used, it just makes the code harder to read.