Fixed a check for non-zero weight when allocating rewards
We expect the pool weight to be non-zero so that we do not allocate rewards to
an empty pool. The condition checking `currentPoolWeight >= 0` was incorrect
given `== 0` case is exactly what we want to avoid. The transaction would fail
later on `totalAmount / currentPoolWeight` but human-readable revert was the
intention of this `require` check.